requirements.txt
openai==1.99.6
requests==2.32.4
python-dotenv==1.1.1
pydantic==2.11.7
